==> adult_vs_pediatric_teaseq/dash_app/01_GetHallmarkGenesetDescriptions.ipynb <==
# Get Hallmark Geneset Descriptions

Save Hallmark geneset .gmt file (v7.5.1) to read into python  
Obtain geneset descriptions via msigdbr 

## Set Up

# install.packages("msigdbr")

library(msigdbr)
library(dplyr)

## Download GMT file

out_gmt <- './data/h.all.v7.5.1.symbols.gmt'

download.file(
    url = 'https://data.broadinstitute.org/gsea-msigdb/msigdb/release/7.5.1/h.all.v7.5.1.symbols.gmt',
    destfile = out_gmt)   

## Format Pathway Descriptions

hallmark_anno <- msigdbr(species = "Homo sapiens", category = "H")

names(hallmark_anno)

head(hallmark_anno, 1)

hallmark_desc <- hallmark_anno %>%
    select('gs_name', 'gs_description') %>%
    distinct()

dim(hallmark_desc)

hallmark_desc

write.csv(hallmark_desc, "data/hallmark_gs_descriptions.csv")

## QC Check GMT vs Query

# Read GMT file in as list of gene sets
# Based on code form GSEABase
readGMT <- function(fp, sep = "\t", ...){
    assertthat::assert_that(file.exists(fp), 
                           msg = sprintf("Could not locate input file %s", fp))
    assertthat::assert_that(grepl(".gmt$", fp),
                            msg = sprintf("Expecting file extension '.gmt'. Input file: %s", fp))
    
    lines <- strsplit(readLines(fp, ...), sep)
    gene_list <- lapply(lines, function(line) {
        unlist(line[-(1:2)])
    })
    names(gene_list) <- sapply(lines, "[[",1)
    gene_list                    
}

# read in all gene sets and merge into one master list
hallmark_gs <- readGMT(out_gmt)

all(hallmark_desc$gs_name %in% names(hallmark_gs))

all(names(hallmark_gs) %in% hallmark_desc$gs_name)

## Session Info

sessionInfo()
