==> adult_vs_pediatric_teaseq/dash_app/02_FormatAppDat_updated.ipynb <==
## Format App Data

Format data for direct App ingest:
* Results data table
* Plot color conventions
* Geneset dictionary

## Set Up

# pip install session_info

import pandas as pd
from warnings import warn
import numpy as np
import hisepy as hp
import os
import session_info
import pickle
import re

## Functions

def check_cat_dict(cat_dict, df, verbose = False):
    ''' 
    Checks that a dictionary of column names and ordered expected values represents complete 
    data given the data frame.
    Requires: from warnings import warn
    '''
    
    for name in cat_dict.keys():
        if(name not in df.columns):
            warn('category dictionary column name not found in data: ' + name)
        else:
            data_cat_all = pd.unique(df[name])
            missing_cat = [x for x in data_cat_all if x not in cat_dict[name]]
            if len(missing_cat) > 0:
                warn('The following values in data column "' + name + '" are not found in category dictionary: ' + ", ".join(missing_cat) + 
                     ". Add to dictionary if you wish to include as categorical levels for plot filtering, etc.")
            extra_cat = [x for x in cat_dict[name] if x not in data_cat_all]
            if len(extra_cat) > 0:
                warn('Extra data levels found in data dictionary for column "' + name + '": ' + ", ".join(extra_cat) + 
                         ". These will be removed as categories during variable formatting.")
    if(verbose):
        print("Finished checking category dictionary")

# #test1: bad dictionary colname
# test1_dict = {'cell_type':ct_order,
#            'randomcol': [1,2,4,23,8],
#           'experiment':expt_order}
# check_cat_dict(test1_dict, df, verbose = True)

# #test2: missing data level
# test2_dict = {'cell_type': ct_order[0:10],
#               'experiment':expt_order}
# check_cat_dict(test2_dict, df, verbose = True)


# #test3: bad dictionary colname and missing data
# test3_dict = {'cell_type':ct_order[0:10],
#                'randomcol': [1,2,4,23,8],
#               'experiment':expt_order}
# check_cat_dict(test3_dict, df, verbose = True)

# #test4: extra category
# test_order = ct_order[0:10].copy()
# test_order.extend(['random_celltype'])
# test4_dict = {'cell_type': test_order,
#               'randomcol': [1,2,4,23,8],
#               'experiment':expt_order}
# check_cat_dict(test4_dict, df, verbose = True)

# check_cat_dict(test_dict, test_df, verbose=True)

def format_df(df, cat_dict={}):
    ''' 
    Converts all string columns to categorical, using input dictionary to set 
    value order if supplied, otherwise sets order as alphanumeric sorted values
    
    Requires: import pandas as pd; import numpy as np
    '''
    res = df.copy()
    
    # Formats
    types = [res[x].dtype for x in res.columns]
    
    # Convert string to categorical
    i_str = [i for i in range(len(types)) if types[i] =='O']
    if(len(i_str) > 0):
        if(len(i_str) > 0):
            cnames = [res.columns[i] for i in i_str]
            cat_dict_all = {x:cat_dict[x] if x in cat_dict.keys() else np.sort(pd.unique(res[x])) for x in cnames}
            for col in cnames:
                res[col] = res[col].astype(pd.CategoricalDtype(categories=cat_dict_all[col], ordered = True))
    return res

def make_comp_label(treatment, timepoint):
    res_str = "TEA-seq " + treatment[0].upper() + treatment[1:3] + ". " + str(int(timepoint)) + " hr"
    return(res_str)


## Read Data

updated data 7/1/2024:

# Pseudobulk expression
pb_uuid = '8e247355-10ef-410c-b031-6512474cf6e8'
hp.cache_files([pb_uuid])
os.rename('/home/jupyter/cache/{u}/ped_sr_tea_pseudobulk_mean_2024-07-01.h5'.format(u = pb_uuid),
          'data/ped_sr_tea_pseudobulk_mean_2024-07-01.h5')

# DEG Results
fid ='fc4fdbfe-6276-4f08-a387-343525b45e26'
fd = hp.read_files(file_list = [fid])
os.rename("cache/unknown/all_mast_deg_2024-07-01.csv", "data/all_mast_deg_2024-07-01.csv")

fp = './data/all_mast_deg_2024-07-01.csv'
df = pd.read_csv(fp)
df.head()

df.shape

## Format Data

### Add treatment column

df['treatment'] = df.loc[:,'fg']

### Add comparison column

df['comparison'] = df.apply(lambda row: 'Older Adult vs Pediatric', axis = 1)

df['comparison'].unique()

### Add celltype label column

ct_df = {
    'aifi_cell_type': [
        't_cd4_naive',
        't_cd4_cm',
        't_cd4_em',
        't_cd4_treg',
        't_cd8_naive',
        't_cd8_memory'
    ],
    'cell_type': [
        'CD4 Naive',
        'CD4 CM',
        'CD4 EM',
        'CD4 Treg',
        'CD8 Naive',
        'CD8 Memory'
    ]
}
ct_labs = pd.DataFrame(ct_df)

df = df.merge(ct_labs.loc[:,['aifi_cell_type','cell_type']], on='aifi_cell_type', how = 'left', copy=True)

df['cell_type'].unique()

### Check

df.shape

df.columns

df.head()

df.tail()

### Variable Levels

Make Categorical data dictionary to explicitly level categorical data. run the `check_cat_dict()` function to check for potential errors

ct_order = ['CD4 Naive','CD4 CM','CD4 EM','CD4 Treg',
            'CD8 Naive', 'CD8 Memory']
comp_order = ['Older Adult vs Pediatric']

cat_dict={'cell_type':ct_order,
          'comparison':comp_order}

check_cat_dict(cat_dict, df, verbose = True)

%%time
df = format_df(df, cat_dict)

### Save Formatted Data

handle_col = open('data/all_mast_deg_2024-07-01.pkl', 'wb')
pickle.dump(df, handle_col, protocol=pickle.HIGHEST_PROTOCOL)
handle_col.close()

## Color dictionary

color_dict_ct = {
    'CD4 Naive': '#93A7D1',
    'CD4 CM': '#00AEEF',
    'CD4 EM': '#1C75BC',
    'CD4 Treg': '#E57A93',
    'CD8 Naive': '#8DC63F',
    'CD8 Memory': '#009444'
} 

comp_dict = {
    'Older Adult vs Pediatric': '#4F8886'
}

preset_colors_dict = {'cell_type':color_dict_ct,
                     'comparison':comp_dict}
preset_colors_dict

### Save color dictionary

handle_col = open('data/meta_color_dict.pkl', 'wb')
pickle.dump(preset_colors_dict, handle_col, protocol=pickle.HIGHEST_PROTOCOL)
handle_col.close()

## Genesets

### Hallmark

# require import re
def fmt_gmt(gmt_fp):
    with open(gmt_fp) as gmt:
        gs_ls = gmt.readlines()
    elem = [x.split("\t") for x in gs_ls]
    gs_names = [x[0] for x in elem]
    gs_list = [x[2:] for x in elem]
    gs_list = [[re.sub("\n$", "", x) for x in gs] for gs in gs_list]
    gs_dict = dict(zip(gs_names, gs_list))
    return(gs_dict)

def add_description_gmt(gs_dict, description = None):
    names = [*gs_dict.keys()]
    gs_list = [*gs_dict.values()]
    if description is None:
        description_list = np.repeat(description, len(names))
    else:
        description_list = description
    res_dict = {names[i]: {'description':description_list[i], 'features':gs_list[i]} for i in range(len(names))}
    return res_dict

gs_description = pd.read_csv("./data/hallmark_gs_descriptions.csv")

desc_dict = dict(zip(gs_description['gs_name'], gs_description['gs_description']))

# read in all gene sets and merge into one master list
hallmark_fp = "./data/h.all.v7.5.1.symbols.gmt"
all_gs =  fmt_gmt(hallmark_fp)
desc_list = [desc_dict[x] for x in all_gs.keys()]
all_gs = add_description_gmt(all_gs, desc_list)

### Aggregate

all_gs.keys()

### Save pathway dictionary

handle_col = open('data/custom_gs_dict.pkl', 'wb')
pickle.dump(all_gs, handle_col, protocol=pickle.HIGHEST_PROTOCOL)
handle_col.close()

## Environment

session_info.show()
