==> adult_vs_pediatric_teaseq/dash_app/03_UploadVisualization.ipynb <==
# Save app to HISE

Save app back to HISE study space

Ensure python version is 3.9 or above for sklearn compatibility with pip-tools

## Set Up

!python --version

import sys 
import os
#sys.path.insert(0, '/home/jupyter/hisepy/build/lib') # Note: reinstalled hisepy into python 3.9 conda environment. adding to path here
import hisepy as hp
import session_info
import glob
import shutil

## Variables

# Destination Study space
spaceid = '00a53fa5-18da-4333-84cb-3cc0b0761201'

# Source DEG dataset
input_fileids = [
    'fc4fdbfe-6276-4f08-a387-343525b45e26', # DEG data
    '8e247355-10ef-410c-b031-6512474cf6e8' # Pseudobulk Data
] 

## Data

# Data files
all_data = [
    'data/ped_sr_tea_pseudobulk_mean_2024-07-01.h5', 
    'data/all_mast_deg_2024-07-01.pkl', 
    'data/custom_gs_dict.pkl', 
    'data/meta_color_dict.pkl',
    'data/app_info.txt', 
    'assets/styles.css']
print(all_data) 
len(all_data)

sizes = [os.stat(x).st_size for x in all_data]
print("{} : {}GB".format("Total Size", round(sum(sizes)/(1024*1024*1024),3)))

for fp in all_data:
    print(fp + " exists: " + str(os.path.exists(fp)))

## Stage files to /home/jupyter/\<app>

target_path = '/home/jupyter/certpro-tea-deg/'
target_data = target_path + 'data/'
target_assets = target_path + 'assets/'

os.mkdir(target_path)
os.mkdir(target_data)
os.mkdir(target_assets)

app_file = 'app.py'
hero_file = 'hero_image.png'

shutil.copy(app_file, target_path + app_file)
for f in all_data:
    shutil.copy(f, target_path + f)
shutil.copy(hero_file, target_path + hero_file)

data_files = [target_path + f for f in all_data]

data_files

target_path + app_file

## Save App

hp.save_dash_app(
    app_filepath = target_path + app_file,
    additional_files = data_files,
    input_file_ids= input_fileids,
    study_space_id = spaceid,
    title = "Pediatric vs Older Adult T cell DEG Explorer, v1.0.3",
    description = "Explore DEGs for between Children and Older Adults in T cell types",
    image = hero_file,
    additional_dirs = []
)

import session_info
session_info.show()
